[MOV] html_builder, *: move the ContentEditablePlugin in html_editor
*: html_editor

The goal of this commit is to move the `ContentEditablePlugin` plugin in
`html_editor` as it is responsible of the `contenteditable` attribute.

Related to task-4367641

X-original-commit: https://github.com/odoo/odoo/commit/0ef47b25f1599fbcb1cfb828da59d3000a34dc45
